+++ ninja-core/src/test/java/ninja/i18n/MessagesImplTest.java
@@ -35,6 +35,8 @@ import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
 
+import com.google.common.base.Optional;
+
 @RunWith(MockitoJUnitRunner.class)
 public class MessagesImplTest {
 
@@ -55,16 +57,16 @@ public class MessagesImplTest {
         Messages lang = new MessagesImpl(ninjaProperties, null);
 
         // that will refer to messages_en.properties:
-        assertEquals("english", lang.get("language", "en-US"));
-        assertEquals("english", lang.get("language", "en-CA"));
-        assertEquals("english", lang.get("language", "en-UK"));
+        assertEquals("english", lang.get("language", Optional.of("en-US")).get());
+        assertEquals("english", lang.get("language", Optional.of("en-CA")).get());
+        assertEquals("english", lang.get("language", Optional.of("en-UK")).get());
 
         // that will refer to messages_de.properties:
-        assertEquals("deutsch", lang.get("language", "de"));
-        assertEquals("deutsch", lang.get("language", "de-DE"));
+        assertEquals("deutsch", lang.get("language", Optional.of("de")).get());
+        assertEquals("deutsch", lang.get("language", Optional.of("de-DE")).get());
 
         // that will refer to messages_fr-FR.properties:
-        assertEquals("francaise", lang.get("language", "fr-FR"));
+        assertEquals("francaise", lang.get("language", Optional.of("fr-FR")).get());
 
     }
     
@@ -82,28 +84,28 @@ public class MessagesImplTest {
 
         // test with context Accept Header
         when(context.getAcceptLanguage()).thenReturn("en-US");
-        assertEquals("english", messages.get("language", context, result));
+        assertEquals("english", messages.get("language", context, Optional.of(result)).get());
         when(context.getAcceptLanguage()).thenReturn("en-CA");
-        assertEquals("english", messages.get("language", context, result));
+        assertEquals("english", messages.get("language", context, Optional.of(result)).get());
         when(context.getAcceptLanguage()).thenReturn("en-UK");
-        assertEquals("english", messages.get("language", context, result));
+        assertEquals("english", messages.get("language", context, Optional.of(result)).get());
 
         // test that result overwrites context AcceptHeader
         lang.setLanguage("de", result);
-        assertEquals("deutsch", messages.get("language", context, result));
+        assertEquals("deutsch", messages.get("language", context, Optional.of(result)).get());
         result = Results.ok();
         lang.setLanguage("de-DE", result);
-        assertEquals("deutsch", messages.get("language", context, result));
+        assertEquals("deutsch", messages.get("language", context, Optional.of(result)).get());
 
         // that forced language from context works with empty result       
         result = Results.ok();
         when(context.getCookie(Mockito.anyString())).thenReturn(
                 Cookie.builder("name", "fr-FR").build());
-        assertEquals("francaise", messages.get("language", context, result));
+        assertEquals("francaise", messages.get("language", context, Optional.of(result)).get());
         //and the result overwrites it again...        
         result = Results.ok();
         lang.setLanguage("de-DE", result);
-        assertEquals("deutsch", messages.get("language", context, result));
+        assertEquals("deutsch", messages.get("language", context, Optional.of(result)).get());
         
 
     }
@@ -117,12 +119,12 @@ public class MessagesImplTest {
         Messages lang = new MessagesImpl(ninjaProperties, null);
 
         // that will refer to messages_fr-FR.properties:
-        assertEquals("francaise", lang.get("language", "da,fr-FR;q=0.8"));
-        assertEquals("francaise", lang.get("language", "da;q=0.9, fr-FR; q=0.8"));
+        assertEquals("francaise", lang.get("language", Optional.of("da,fr-FR;q=0.8")).get());
+        assertEquals("francaise", lang.get("language", Optional.of("da;q=0.9, fr-FR; q=0.8")).get());
         
         // that will refer to messages_de.properties:
-        assertEquals("deutsch", lang.get("language", "de,fr-FR;q=0.8"));
-        assertEquals("deutsch", lang.get("language", "de;q=0.9, fr-FR; q=0.8"));
+        assertEquals("deutsch", lang.get("language", Optional.of("de,fr-FR;q=0.8")).get());
+        assertEquals("deutsch", lang.get("language", Optional.of("de;q=0.9, fr-FR; q=0.8")).get());
 
     }
 
@@ -135,17 +137,17 @@ public class MessagesImplTest {
 
         // that will refer to messages_en.properties:
         assertEquals("this is the placeholder: test_parameter",
-                lang.get("message_with_placeholder", "en-US", "test_parameter"));
+                lang.get("message_with_placeholder", Optional.of("en-US"), "test_parameter").get());
         assertEquals("this is the placeholder: test_parameter",
-                lang.get("message_with_placeholder", "en-CA", "test_parameter"));
+                lang.get("message_with_placeholder", Optional.of("en-CA"), "test_parameter").get());
         assertEquals("this is the placeholder: test_parameter",
-                lang.get("message_with_placeholder", "en-UK", "test_parameter"));
+                lang.get("message_with_placeholder", Optional.of("en-UK"), "test_parameter").get());
 
         // that will refer to messages_de.properties:
         assertEquals("das ist der platzhalter: test_parameter",
-                lang.get("message_with_placeholder", "de", "test_parameter"));
+                lang.get("message_with_placeholder", Optional.of("de"), "test_parameter").get());
         assertEquals("das ist der platzhalter: test_parameter",
-                lang.get("message_with_placeholder", "de-DE", "test_parameter"));
+                lang.get("message_with_placeholder", Optional.of("de-DE"), "test_parameter").get());
 
     }
     
@@ -165,34 +167,34 @@ public class MessagesImplTest {
 
         // that will refer to messages_en.properties:
         assertEquals("this is the placeholder: test_parameter",
-                messages.get("message_with_placeholder", context, result, "test_parameter"));       
+                messages.get("message_with_placeholder", context, Optional.of(result), "test_parameter").get());       
         when(context.getAcceptLanguage()).thenReturn("en-CA");
         assertEquals("this is the placeholder: test_parameter",
-                messages.get("message_with_placeholder", context, result, "test_parameter"));
+                messages.get("message_with_placeholder", context, Optional.of(result), "test_parameter").get());
         when(context.getAcceptLanguage()).thenReturn("en-UK");
         assertEquals("this is the placeholder: test_parameter",
-                messages.get("message_with_placeholder", context, result, "test_parameter"));
+                messages.get("message_with_placeholder", context, Optional.of(result), "test_parameter").get());
 
         
         // that will refer to messages_de.properties:
         lang.setLanguage("de", result);
         assertEquals("das ist der platzhalter: test_parameter",
-                messages.get("message_with_placeholder", context, result, "test_parameter"));
+                messages.get("message_with_placeholder", context, Optional.of(result), "test_parameter").get());
         
         lang.setLanguage("de-DE", result);
         assertEquals("das ist der platzhalter: test_parameter",
-                messages.get("message_with_placeholder", context, result, "test_parameter"));
+                messages.get("message_with_placeholder", context, Optional.of(result), "test_parameter").get());
 
         
         // that forced language from context works with empty result       
         result = Results.ok();
         when(context.getCookie(Mockito.anyString())).thenReturn(
                 Cookie.builder("name", "fr-FR").build());
-        assertEquals("cest le placeholder: test_parameter", messages.get("message_with_placeholder", context, result, "test_parameter"));
+        assertEquals("cest le placeholder: test_parameter", messages.get("message_with_placeholder", context, Optional.of(result), "test_parameter").get());
         //and the result overwrites it again...        
         result = Results.ok();
         lang.setLanguage("de-DE", result);
-        assertEquals("das ist der platzhalter: test_parameter", messages.get("message_with_placeholder", context, result, "test_parameter"));
+        assertEquals("das ist der platzhalter: test_parameter", messages.get("message_with_placeholder", context, Optional.of(result), "test_parameter").get());
     }
     
     
@@ -205,7 +207,7 @@ public class MessagesImplTest {
         Messages lang = new MessagesImpl(ninjaProperties, null);
         
         // US locale testing:
-        Map<Object, Object> map = lang.getAll("en-US");
+        Map<Object, Object> map = lang.getAll(Optional.of("en-US"));
 
         assertEquals(4, map.keySet().size());
         assertTrue(map.containsKey("language"));
@@ -216,7 +218,7 @@ public class MessagesImplTest {
         assertEquals("english", map.get("language"));
 
         // GERMAN locale testing:
-        map = lang.getAll("de");
+        map = lang.getAll(Optional.of("de"));
         assertEquals(4, map.keySet().size());
         assertTrue(map.containsKey("language"));
         assertTrue(map.containsKey("message_with_placeholder"));
@@ -243,7 +245,7 @@ public class MessagesImplTest {
         when(context.getAcceptLanguage()).thenReturn("en-US");
         
         // US locale testing:
-        Map<Object, Object> map = messages.getAll(context, result);
+        Map<Object, Object> map = messages.getAll(context, Optional.of(result));
 
         assertEquals(4, map.keySet().size());
         assertTrue(map.containsKey("language"));
@@ -256,7 +258,7 @@ public class MessagesImplTest {
         // GERMAN locale testing:
         lang.setLanguage("de", result);
         
-        map = messages.getAll(context, result);
+        map = messages.getAll(context, Optional.of(result));
         assertEquals(4, map.keySet().size());
         assertTrue(map.containsKey("language"));
         assertTrue(map.containsKey("message_with_placeholder"));
@@ -272,7 +274,7 @@ public class MessagesImplTest {
         result = Results.ok();
         when(context.getCookie(Mockito.anyString())).thenReturn(
                 Cookie.builder("name", "en").build());
-        map = messages.getAll(context, result);
+        map = messages.getAll(context, Optional.of(result));
 
         assertEquals(4, map.keySet().size());
         assertTrue(map.containsKey("language"));
@@ -295,7 +297,7 @@ public class MessagesImplTest {
         Messages lang = new MessagesImpl(ninjaProperties, null);
 
         assertEquals("prop1, prop2, prop3",
-                lang.get("a_propert_with_commas", "en-US"));
+                lang.get("a_propert_with_commas", Optional.of("en-US")).get());
 
     }